Adding "-PJ n" in GEP option for profiled jumps : enable a switch of n instructions inside the interpretor loop. The switch permits better prediction branch than the usual indirect jump to an instruction handler.
New template symbole : $(max_operand_nb) which gives the maximum number of operands for all the given set of instructions.

Adding option : GLISS_DCACHE_WITH_INSTR_WORD in GEP.
When -on decode hashtable for the cache is based on the opcode of the instruction rather than his address

Adding option : -b -bench. In the simulator -> the programm loops until the execution time exceded 200ms